patterns/01_structural/04_decorator/decorator_test.go
// Acceptance criteria
//
// The acceptance criteria for a Decorator pattern is to have a common interface
// and a core type, the one that all layers will be built over:
//
// We must have the main interface that all decorators will implement. This
// interface will be called IngredientAdd, and it will have the AddIngredient()
// string method.
//
// We must have a core PizzaDecorator type (the decorator) that we will add
// ingredients to.
//
// We must have an ingredient "onion" implementing the same IngredientAdd
// interface that will add the string onion to the returned pizza.
//
// We must have a ingredient "meat" implementing the IngredientAdd interface
// that will add the string meat to the returned pizza.
//
// When calling AddIngredient method on the top object, it must return a fully
// decorated pizza with the text Pizza with the following ingredients: meat,
// onion.
//
package main

import (
	"strings"
	"testing"
)

func TestPizzaDecorator_AddIngredient(t *testing.T) {
	pizza := &PizzaDecorator{}
	pizzaResult, _ := pizza.AddIngredient()
	expectedText := "Pizza with the following ingredients:"

	if !strings.Contains(pizzaResult, expectedText) {
		t.Errorf("When calling the add ingredient of the pizza decorator it must "+
			"return the text %sthe expected text, not '%s'", pizzaResult, expectedText)
	}

}

func TestOnion_AddIngredient(t *testing.T) {
	onion := &Onion{}
	onionResult, err := onion.AddIngredient()
	if err == nil {
		t.Errorf("When calling AddIngredient on the onion decorator without an "+
			"IngredientAdd on its Ingredient field must return an error, not a "+
			"string with '%s'", onionResult)
	}

	onion = &Onion{&PizzaDecorator{}}
	onionResult, err = onion.AddIngredient()

	if err != nil {
		t.Error(err)
	}

	if !strings.Contains(onionResult, "onion") {
		t.Errorf("When calling the add ingredient of the onion decorator it must "+
			"return a text with the word 'onion', not '%s'", onionResult)
	}
}

func TestMeat_AddIngredient(t *testing.T) {
	meat := &Meat{}
	meatResult, err := meat.AddIngredient()
	if err == nil {
		t.Errorf("When calling AddIngredient on the meat decorator without"+
			"an IngredientAdd in its Ingredient field must return an error,"+
			"not a string with '%s'", meatResult)
	}

	meat = &Meat{&PizzaDecorator{}}
	meatResult, err = meat.AddIngredient()

	if err != nil {
		t.Error(err)
	}

	if !strings.Contains(meatResult, "meat") {
		t.Errorf("When calling the add ingredient of the meat decorator it"+
			"must return a text with the word 'meat', not '%s'", meatResult)
	}
}

func TestPizzaDecorator_FullStack(t *testing.T) {
	pizza := &Onion{&Meat{&PizzaDecorator{}}}
	pizzaResult, err := pizza.AddIngredient()
	if err != nil {
		t.Error(err)
	}

	expectedText := "Pizza with the following ingredients: meat, onion"

	if !strings.Contains(pizzaResult, expectedText) {
		t.Errorf("When asking for a pizza with onion and meat the returned "+
			"string must contain the text '%s' but '%s' didn't have it",
			expectedText, pizzaResult)
	}

	t.Log(pizzaResult)
}
